Multiple Selectors
Attempted to create a selector component in order to find the buildingTypes associated with a council. 

So far able to have one component on the homepage which chooses and prints the council name in console.

Second selector is a reusable component that attempts to take in specific arrays so that its flexible. Sorta works. But need to refine.