Multiple Pages
Able to enter building types on a separate stage. This updates the state.

Created a required drop select, however its unable to be detected when not input. Also chosen selection doesn't update state as well... This makes it very tricky and thus manual when choosing and updating results to the next stage of the form